requirements.txt
Autologging==1.0.1
Flask==0.10.1
Jinja2==2.8
Werkzeug==0.11.8
requests==2.9.1
Flask-Compress==1.4.0
flask-sqlalchemy==2.4.0
flask-compress==1.4.0
flask-cors==3.0.7
base==1.0.4